requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kiwisolver==1.3.1
langdetect==1.0.8
MarkupSafe==1.1.1
matplotlib==3.2.0
nltk==3.6.2
numpy==1.20.2
oauthlib==3.1.0
pandas==1.2.4
Pillow==8.2.0
praw==7.2.0
prawcore==2.0.0
pycountry==20.7.3
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
sklearn==0.0
textblob==0.15.3
threadpoolctl==2.1.0
tqdm==4.60.0
tweepy==3.10.0
update-checker==0.18.0
urllib3==1.26.4
websocket-client==0.59.0
Werkzeug==1.0.1
